refactor(organizationstructureguard): moved guard to main api folder
